Fixing deletion. Was not able to convert line with spaces to new format.

+++ deleteFile.py
@@ -19,6 +19,9 @@ class DeleteFileCommand(sublime_plugin.TextCommand):
 		if cT.projectFileExistence(info.getCurrentProject(), info.getCurrentProjectPath()):
 			print("Project file not found or it is empty.")
 			return 0
+		if not cT.fileExistence(name, info.getCurrentProjectPath()):
+			print("File to delete not found.")
+			return 0
 
 		projectWriter = rw.ProjectWriter()
 		if projectWriter.deleteFile(name):
